+++ libs/efw/src/output_ctl.rs
@@ -20,7 +20,7 @@ impl<'a> OutputCtl {
     const COEF_MIN: i32 = 0x00000000;
     const COEF_MAX: i32 = 0x02000000;
     const COEF_STEP: i32 = 0x00000001;
-    const COEF_TLV: &'a [i32] = &[5, 8, -14400, 6];
+    const COEF_TLV: &'a [u32] = &[5, 8, -14400i32 as u32, 6];
 
     const OUT_NOMINAL_LABELS: &'a [&'a str] = &["+4dBu", "-10dBV"];
     const OUT_NOMINAL_LEVELS: &'a [NominalLevel] = &[NominalLevel::PlusFour, NominalLevel::MinusTen];
